Add program to zip working directory iteratively.
This program zips all files and subdirectories of the cwd and
names the files iteratively – i.e. dir_1.zip, dir_2.zip, and so on.
Any new zip does NOT include the previous zipped files, such that
we do not include e.g. dir_2.zip in dir_3.zip. The compressed directory
is placed in the cwd.